tagged_base64

Attribute Macro tagged

#[tagged]

Derive serdes for a type which serializes as a binary blob.

This macro can be used to easily derive friendly serde implementations for a binary type which implements CanonicalSerialize and CanonicalDeserialize. This is useful for cryptographic primitives and other types which do not have a human-readable serialization, but which may be embedded in structs with a human-readable serialization. The serde implementations derived by this macro will serialize the type as bytes for binary encodings and as base 64 for human readable encodings.

This macro takes at least one arguments:

The first argument should be the tag, as a string literal or expression.

By default, the derived implementation invokes CanonicalSerialize and CanonicalDeserialize with uncompressed and unchecked flags.

If compressed and/or checked flags are presented, the derived implementation will behave accordingly.

Specifically, this macro does 4 things when applied to a type definition:

It adds #[derive(Serialize, Deserialize)] to the type definition, along with serde attributes to serialize using TaggedBase64.

It creates an implementation of Tagged for the type using the specified tag. This tag will be used to identify base 64 strings which represent this type in human-readable encodings.

It creates an implementation of TryFrom<TaggedBase64> for the type T, which is needed to make the serde(try_from) attribute work.

It creates implementations of Display and FromStr using tagged base 64 as a display format. This allows tagged blob types to be conveniently displayed and read to and from user interfaces in a manner consistent with how they are serialized.

Usage example:

use ark_serialize::*;
use tagged_base64_macros::tagged;

#[tagged("PRIM")]
#[derive(Clone, CanonicalSerialize, CanonicalDeserialize, /* any other derives */)]
pub struct CryptoPrim(
    // This type can only be serialied as an opaque, binary blob using ark_serialize.
    pub(crate) ark_bls12_381::Fr,
);

The type CryptoPrim can now be serialized as binary:

bincode::serialize(&crypto_prim).unwrap();

or as base64:

serde_json::to_string(&crypto_prim).unwrap();

which will produce a tagged base64 string like “PRIM~8oaujwbov8h4eEq7HFpqW6mIXhVbtJGxLUgiKrGpMCoJ”.